refactor: 스토리북 삭제, 사이프레스 주석
