+++ src/router/pages/NotFound/Test.tsx
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+const TestPage = () => {
+  return <TestStyled></TestStyled>;
+};
+
+const TestStyled = styled.div``;
+
+export default TestPage;
