--- src/stories/texts/FooterContactItem.stories.tsx
@@ -1,39 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react';
-
-import FooterContactItem from '@router/layout/Footer/components/FooterContactItem';
-
-const meta: Meta<typeof FooterContactItem> = {
-  title: 'Texts/FooterContactItem',
-  component: FooterContactItem,
-  parameters: {
-    layout: 'centered',
-  },
-  tags: ['autodocs'],
-  decorators: [
-    (Story) => (
-      <ul>
-        <Story />
-      </ul>
-    ),
-  ],
-  argTypes: {
-    children: {
-      control: 'text',
-      description: 'FooterContactItem 정보',
-    },
-    title: {
-      control: 'text',
-      description: 'FooterContactItem 테이틀',
-    },
-  },
-};
-
-export default meta;
-type Story = StoryObj<typeof meta>;
-
-export const Default: Story = {
-  args: {
-    children: 'ckehfqkr302@gamil.com',
-    title: 'General Enquiry',
-  },
-};
