--- src/stories/texts/ProductCardInfo.stories.tsx
@@ -1,34 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react';
-
-import ProductCardInfo from '@router/pages/categories/products/components/ProductCardInfo';
-
-const meta: Meta<typeof ProductCardInfo> = {
-  title: 'Texts/ProductCardInfo',
-  component: ProductCardInfo,
-  parameters: {
-    layout: 'centered',
-  },
-  tags: ['autodocs'],
-  decorators: [
-    (Story) => (
-      <ul>
-        <Story />
-      </ul>
-    ),
-  ],
-  argTypes: {
-    children: {
-      control: 'text',
-      description: 'ProductCardInfo 텍스트',
-    },
-  },
-};
-
-export default meta;
-type Story = StoryObj<typeof meta>;
-
-export const Default: Story = {
-  args: {
-    children: '₩ 140,000',
-  },
-};
