--- src/stories/texts/ProductInfo.stories.tsx
@@ -1,44 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react';
-
-import ProductInfo from '@router/pages/categories/products/components/ProductInfo';
-
-const meta: Meta<typeof ProductInfo> = {
-  title: 'Texts/ProductInfo',
-  component: ProductInfo,
-  parameters: {
-    layout: 'centered',
-  },
-  tags: ['autodocs'],
-  decorators: [
-    (Story) => (
-      <div style={{ width: '600px' }}>
-        <Story />
-      </div>
-    ),
-  ],
-  argTypes: {
-    children: {
-      control: 'text',
-      description: 'ProductInfo 상품 정보',
-    },
-    name: {
-      control: 'text',
-      description: 'ProductInfo 상품 이름',
-    },
-    price: {
-      control: 'number',
-      description: 'ProductInfo 상품 가격',
-    },
-  },
-};
-
-export default meta;
-type Story = StoryObj<typeof meta>;
-
-export const Default: Story = {
-  args: {
-    children: 'Product infos',
-    price: 95000,
-    name: 'Product Name',
-  },
-};
